# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/register": {
      "post": {
        "operationId": "UsersController_create",
        "summary": "注册新用户",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "用户创建成功"
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/users/login": {
      "post": {
        "operationId": "UsersController_login",
        "summary": "用户登录",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功"
          }
        },
        "tags": [
          "users"
        ]
      }
    },
...
    "/users/friend-requests": {
      "post": {
        "operationId": "UsersController_sendFriendRequest",
        "summary": "发送好友请求",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFriendRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "请求发送成功"
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "UsersController_getFriendRequests",
        "summary": "获取发给我的好友请求",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "PENDING",
                "ACCEPTED",
                "REJECTED"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功"
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/users/friend-requests/{id}": {
      "patch": {
        "operationId": "UsersController_handleFriendRequest",
        "summary": "处理好友请求",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFriendRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求处理成功"
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/friends": {
      "get": {
        "operationId": "UsersController_getFriends",
        "summary": "获取好友列表",
        "parameters": [],
        "responses": {
          "200": {
            "description": "获取成功"
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/search": {
      "get": {
        "operationId": "UsersController_searchUsers",
        "summary": "搜索用户",
        "parameters": [
          {
            "name": "keyword",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "搜索成功"
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/users/profile": {
      "get": {
        "operationId": "UsersController_getProfile",
        "summary": "获取当前用户信息",
        "parameters": [],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number"
                        },
                        "username": {
                          "type": "string"
                        },
                        "avatar": {
                          "type": "string"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_getUserInfo",
        "summary": "获取指定用户信息",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "用户ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number"
                        },
                        "username": {
                          "type": "string"
                        },
                        "avatar": {
                          "type": "string"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Chat API",
    "description": "The chat API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "user1",
            "description": "用户名"
          },
          "password": {
            "type": "string",
            "example": "encrypted_password",
            "description": "加密后的密码"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "user1",
            "description": "用户名"
          },
          "password": {
            "type": "string",
            "example": "encrypted_password",
            "description": "加密后的密码"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "CreateFriendRequestDto": {
        "type": "object",
        "properties": {
          "toId": {
            "type": "number",
            "description": "接收者ID"
          },
          "message": {
            "type": "string",
            "description": "请求消息"
          }
        },
        "required": [
          "toId"
        ]
      },
      "UpdateFriendRequestDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "ACCEPTED",
              "REJECTED"
            ],
            "description": "请求状态"
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}